User/Authentication: we can now fully log into the system.
User: set roles.owner=this.
UserEndpoint: handle stale session.
UserService: implemented findById().
Role: made id @GeneratedValue, added Builder.
BootstrappingDataProviderSingleton: removed old mock data, now creates proper User->Role->Employment, mapped to OrgUnit bootstrap data.

bootstrap-organization.json: added a couple of email adresses fro lookup purposes.

+++ src/main/java/dk/os2opgavefordeler/service/UserService.java
@@ -13,7 +13,7 @@ import java.util.Optional;
  * @author hlo@miracle.dk
  */
 public interface UserService {
-	Optional<User> findById(int userId);
+	Optional<User> findById(long userId);
 	Optional<User> findByEmail(String email);
 	User createUser(User user);
